Adding success after order placed

+++ database/migrations/2022_01_30_094954_create_order_items_table.php
@@ -15,7 +15,7 @@ class CreateOrderItemsTable extends Migration
     {
         Schema::create('order_items', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('order_id');
+            $table->unsignedBigInteger('order_id');
             $table->integer('product_id');
             $table->integer('quantity');
             $table->integer('price');
